Webpack config in TS for online-editor
