+++ packages/chrome-extension-pack-kogito-kie-editors/package.json
@@ -45,7 +45,7 @@
     "@types/jest-when": "^2.7.4",
     "@types/selenium-webdriver": "^4.1.7",
     "chai": "^4.3.4",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
     "jest-when": "^3.5.0",
@@ -55,9 +55,9 @@
     "ts-jest": "^26.5.6",
     "typescript": "^4.4.2",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3",
+    "webpack-merge": "^5.8.0",
     "zip-webpack-plugin": "^4.0.1"
   }
 }
