+++ packages/chrome-extension-serverless-workflow-editor/package.json
@@ -43,7 +43,7 @@
     "@types/chrome": "^0.0.193",
     "@types/jest": "^26.0.23",
     "@types/jest-when": "^2.7.4",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "jest": "^26.6.3",
     "jest-when": "^3.5.0",
     "monaco-editor": "^0.33.0",
@@ -54,9 +54,9 @@
     "ts-jest": "^26.5.6",
     "typescript": "^4.4.2",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3",
+    "webpack-merge": "^5.8.0",
     "zip-webpack-plugin": "^4.0.1"
   }
 }
