+++ packages/dashbuilder-component-assembler/package.json
@@ -33,11 +33,11 @@
     "@kie-tools-core/webpack-base": "workspace:*",
     "@kie-tools/root-env": "workspace:*",
     "@kie-tools/tsconfig": "workspace:*",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "rimraf": "^3.0.2",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   }
 }
