+++ packages/dashbuilder-component-map/package.json
@@ -47,7 +47,7 @@
     "@types/react-dom": "^17.0.5",
     "@types/react-simple-maps": "^1.0.8",
     "@types/testing-library__jest-dom": "^5.9.1",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "html-webpack-plugin": "^5.3.2",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
@@ -56,8 +56,8 @@
     "ts-jest": "^26.5.6",
     "typescript": "^4.4.2",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   }
 }
