+++ packages/dashbuilder-component-table/package.json
@@ -45,7 +45,7 @@
     "@types/react": "^17.0.6",
     "@types/react-dom": "^17.0.5",
     "@types/testing-library__jest-dom": "^5.9.1",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "copyfiles": "^2.4.1",
     "html-webpack-plugin": "^5.3.2",
     "jest": "^26.6.3",
@@ -55,8 +55,8 @@
     "ts-jest": "^26.5.6",
     "typescript": "^4.4.2",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   }
 }
