+++ packages/online-editor/tsconfig.json
@@ -9,5 +9,15 @@
     "allowSyntheticDefaultImports": true,
     "target": "es6"
   },
-  "include": ["src"]
+  "include": ["src"],
+
+  // Required to make Webpack TS config work. See https://github.com/webpack/webpack-cli/issues/2458#issuecomment-1157987399
+  "ts-node": {
+    "esm": true,
+    "compilerOptions": {
+      "esModuleInterop": true,
+      "module": "CommonJS",
+      "moduleResolution": "NodeNext"
+    }
+  }
 }
