+++ packages/serverless-workflow-text-editor/package.json
@@ -63,7 +63,7 @@
     "@types/json-schema": "^7.0.11",
     "@types/react": "^17.0.6",
     "@types/testing-library__jest-dom": "^5.9.1",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "cypress": "^10.7.0",
     "cypress-file-upload": "^5.0.8",
     "cypress-iframe": "^1.0.1",
@@ -78,8 +78,8 @@
     "typescript": "^4.4.2",
     "vscode-json-languageservice": "^4.2.1",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   }
 }
