+++ packages/uniforms-bootstrap4-codegen/package.json
@@ -48,7 +48,7 @@
     "@types/simpl-schema": "^1.12.0",
     "@types/testing-library__react": "^9.1.2",
     "@types/underscore": "^1.11.2",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
     "jest-raw-loader": "^1.0.1",
@@ -59,7 +59,7 @@
     "typescript": "^4.4.2",
     "webpack": "^5.36.2",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3",
+    "webpack-merge": "^5.8.0",
     "webpack-node-externals": "^3.0.0"
   }
 }
