+++ packages/uniforms-patternfly-codegen/package.json
@@ -46,7 +46,7 @@
     "@types/react-dom": "^17.0.5",
     "@types/simpl-schema": "^1.12.0",
     "@types/testing-library__react": "^9.1.2",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
     "jest-raw-loader": "^1.0.1",
@@ -56,9 +56,9 @@
     "ts-jest": "^26.5.6",
     "typescript": "^4.4.2",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3",
+    "webpack-merge": "^5.8.0",
     "webpack-node-externals": "^3.0.0"
   }
 }
