+++ packages/vscode-extension-backend/package.json
@@ -42,7 +42,7 @@
     "@types/jest": "^26.0.23",
     "@types/jest-when": "^2.7.4",
     "@types/vscode": "^1.66.0",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "jest": "^26.6.3",
     "jest-junit": "^14.0.0",
     "jest-when": "^3.5.0",
@@ -51,9 +51,9 @@
     "ts-jest": "^26.5.6",
     "vsce": "^2.6.5",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   },
   "engines": {
     "vscode": "^1.66.0"
