+++ packages/vscode-extension-red-hat-business-automation-bundle/package.json
@@ -31,9 +31,9 @@
     "rimraf": "^3.0.2",
     "vsce": "^2.6.5",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   },
   "engines": {
     "vscode": "^1.66.0"
