+++ packages/vscode-extension-serverless-workflow-editor/package.json
@@ -63,7 +63,7 @@
     "@types/vscode": "^1.66.0",
     "@vscode/test-web": "^0.0.30",
     "chai": "^4.3.4",
-    "copy-webpack-plugin": "^8.1.1",
+    "copy-webpack-plugin": "^11.0.0",
     "cpr": "^3.0.1",
     "mocha": "^9.2.0",
     "mocha-jenkins-reporter": "^0.4.5",
@@ -77,9 +77,9 @@
     "vscode-extension-tester": "^4.4.0",
     "vscode-test": "^1.6.1",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   },
   "engines": {
     "vscode": "^1.66.0"
