+++ packages/vscode-extension-yard-editor/package.json
@@ -69,9 +69,9 @@
     "vscode-extension-tester": "^4.4.0",
     "vscode-test": "^1.6.1",
     "webpack": "^5.36.2",
-    "webpack-cli": "^4.7.0",
+    "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.7.3",
-    "webpack-merge": "^5.7.3"
+    "webpack-merge": "^5.8.0"
   },
   "engines": {
     "vscode": "^1.66.0"
