+++ packages/webpack-base/package.json
@@ -12,6 +12,7 @@
   "bugs": {
     "url": "https://github.com/kiegroup/kie-tools/issues"
   },
+  "types": "webpack.common.config.d.ts",
   "files": [
     "."
   ],
@@ -21,6 +22,7 @@
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "source-map-loader": "^2.0.2",
-    "ts-loader": "^9.1.2"
+    "ts-loader": "^9.1.2",
+    "webpack": "^5.36.2"
   }
 }
